requirements.txt
amply==0.1.2
appnope==0.1.0
argon2-cffi==20.1.0
attrs==19.3.0
backcall==0.2.0
basketball-reference-web-scraper==4.10.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2018.10.15
cffi==1.14.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
fuzzywuzzy==0.18.0
git-python==1.0.3
gitdb==4.0.5
GitPython==3.1.7
idna==2.7
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
lxml==4.5.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.1
numpy==1.19.1
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
PuLP==2.3
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2018.6
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
